--- a/build.rs
+++ b/build.rs
@@ -92,6 +92,7 @@ fn build_nasm_files() {
     "src/x86/looprestoration_avx2.asm",
     "src/x86/looprestoration16_avx2.asm",
     "src/x86/mc_avx2.asm",
+    "src/x86/mc16_avx2.asm",
     "src/x86/mc_avx512.asm",
     "src/x86/mc_sse.asm",
     "src/x86/me.asm",