--- a/src/asm/x86/mc.rs
+++ b/src/asm/x86/mc.rs
@@ -7,6 +7,8 @@
 // Media Patent License 1.0 was not distributed with this source code in the
 // PATENTS file, you can obtain it at www.aomedia.org/license/patent.
 
+#![allow(dead_code)]
+
 use crate::cpu_features::CpuFeatureLevel;
 use crate::frame::*;
 use crate::mc::FilterMode::*;
@@ -355,7 +357,7 @@ decl_mc_hbd_fns!(
 cpu_function_lookup_table!(
   PUT_HBD_FNS: [[Option<PutHBDFn>; 16]],
   default: [None; 16],
-  [AVX2]
+  []
 );
 
 macro_rules! decl_mct_fns {
@@ -462,7 +464,7 @@ decl_mct_hbd_fns!(
 cpu_function_lookup_table!(
   PREP_HBD_FNS: [[Option<PrepHBDFn>; 16]],
   default: [None; 16],
-  [AVX2]
+  []
 );
 
 extern {
@@ -488,11 +490,7 @@ cpu_function_lookup_table!(
   [(SSSE3, Some(rav1e_avg_ssse3)), (AVX2, Some(rav1e_avg_avx2))]
 );
 
-cpu_function_lookup_table!(
-  AVG_HBD_FNS: [Option<AvgHBDFn>],
-  default: None,
-  [(AVX2, Some(rav1e_avg_16bpc_avx2))]
-);
+cpu_function_lookup_table!(AVG_HBD_FNS: [Option<AvgHBDFn>], default: None, []);
 
 #[cfg(test)]
 mod test {