This implements dav1d's assembly code up to commit 1cf1b309 which adds their 0.9.0 changelog (although not officially tagged yet), which brings some huge improvements for high bit depth on x86_64. This does not pull in any NEON assembly updates between 0.8.2 to 0.9.0.

The following benchmarks were run on a 10-bit, 720p sample clip on a Ryzen 4800H.

|Benchmark|Before FPS|After FPS|Speedup|
|---|---|---|---|
|Speed 3, 10 frames|0.256|0.409|60%|
|Speed 6, 10 frames|0.802|1.335|66%|
|Speed 6, 60 frames|0.853|1.385|62%|
|Speed 10, 60 frames|2.441|4.349|78%|